refact

+++ sources/DLoader/src/Utils/Buffer.cpp
@@ -6,7 +6,7 @@
 
 
 
-Buffer::Buffer(uint s)
+Buffer::Buffer(int s)
 {
     Malloc(s);
 }
@@ -18,7 +18,7 @@ Buffer::~Buffer()
 }
 
 
-void Buffer::Realloc(uint _size)
+void Buffer::Realloc(int _size)
 {
     Free();
     Malloc(_size);
@@ -29,7 +29,7 @@ void Buffer::Fill(uint8 value)
 {
     if (size)
     {
-        std::memset(data, value, size);
+        std::memset(data, value, static_cast<uint>(size));
     }
 }
 
@@ -42,12 +42,12 @@ void Buffer::Free()
 }
 
 
-void Buffer::Malloc(uint s)
+void Buffer::Malloc(int s)
 {
     if (s > 0)
     {
-        data = static_cast<uint8 *>(std::malloc(s));
-        size = (data) ? s : 0U;
+        data = static_cast<uint8 *>(std::malloc(static_cast<uint>(s)));
+        size = (data) ? s : 0;
 
         if(!data)
         {
